Check for VAR_SSO when using topo_wind
TYPE: bug fix

KEYWORDS: VAR_SSO, topo_wind

SOURCE: Internal

DESCRIPTION OF CHANGES: 
Versions of WPS prior to V3.4 did not include VAR_SSO static data. This is necessary for running the topo_wind option in WRF (also introduced in version 3.4). If someone tries to use this option with a newer version of WRF, but with older WPS files, this causes problems, and we need this to cause a fatal error. Added a state variable to Registry.EM_COMMON called 'got_var_sso' (logical). Added a check in module_initialize_real.F to determine if var_sso is zero anywhere in the domain. If so, we set it to .true. Because we need to use a function in the module_initialize_real.F check that returns an 'OR' logical when looping through multiple processors, in module_dm.F, it was necessary to create a logical function called 'wrf_dm_lor_logical' that takes the input value and gives a return value (in the meantime, went ahead and added a similar function for an 'AND' logical in case it's needed in the future). Finally, added a check in start_em.F to make sure if using the topo wind option, VARSSO data is available. If not, it calls a fatal error and prints 'topo_wind requires VAR_SSO data'.

LIST OF MODIFIED FILES: 
M Registry/Registry.EM_COMMON
M dyn_em/module_initialize_real.F
M dyn_em/start_em.F
M external/RSL_LITE/module_dm.F

TESTS CONDUCTED: Ran tests to verify that when using WPS data without VAR_SSO, with the topo_wind option, I get a fatal error specifying the reason. Regression test successful.