bug fix for WRF IO_QUILT (#13)
TYPE: bug fix

KEYWORDS: WRF IO_QUILT

SOURCE: internal (Dave)

DESCRIPTION OF CHANGES: The subroutine call was missing an argument, even though the argument is not used directly. The Cray compiler usually segfaults on this line (the code was wrong, the Cray compiler just indicated the error). If the quilting option was working for a user before this modification, it will continue to work after this modification (consistent with what was found with the regression tests, which do quilting

LIST OF MODIFIED FILES: 
M frame/module_io_quilt_old.F

TESTS CONDUCTED: A nesting run with this fix is done to confirm this fix works. Regression is OK.