Fix DA VARBC.in default scanning pred_use flags for AMSR-2 (#35)
TYPE: text only

KEYWORDS: WRFDA, AMSR-2, VARBC.in

SOURCE: internal

DESCRIPTION OF CHANGES:
The example VARBC.in does not set a good example for AMSR-2.
AMSR-2 (Advanced Microwave Scanning Radiometer 2) is a conical scanning instrument.
Scan-angle bias correction is not necessary and the use_flag should be -1
for scan angle entries in VARBC.in.

The VARBC.in file in the test case that the original developer provided had the
correct -1 settings. The text error was introduced by the code merging proxy.
But turning on the scan-angle bias correction is not really wrong, it is just not
necessary.

LIST OF MODIFIED FILES:
M var/run/VARBC.in

TESTS CONDUCTED:
The AMSR-2 test case result changed due to the different total bias correction applied.
Scan-angle bias term should be small but there will be non-zero value if the use_flag is on.