Splitting the PD advection limiter loop for better performance #46
TYPE: Improvement

KEYWORDS: Advection, Positive Definite, Optimization

SOURCE: Negin Sobhani (UIowa) | Davide Del Vento (NCAR)

DESCRIPTION OF CHANGES:
This is the first of the changes that I worked on with @davidedelvento .
Tests on the advection kernel showed that PD advection limiter loop is not vectorized and is a major hotspot of the subroutine. The PD condition check loop was split and reordered to increase the vectorization for the first two sections loops and accelerate the code.
This change showed a 100% speed-up (2X) for PD condition section of the code (the loop that has been split) and of ~17% speed-up for the advection kernel for intel compiler (-O3). For GNU compiler (-Ofast), we have 105% speed-up on the PD limiter section of the code and total of ~11% speed-up for the advection kernel. For PGI compiler with -O3, we have only 35% speed-up for the PD limiter loop and total of ~ 4% speed-up.
Depending on how much advection you have in your WRF case, this speed-up might be significant.
For example, advection might include up to 50% of the total time of the code and a 17% speed-up for advection kernel means a total of 8.5% speed-up for the whole WRF run. For an advection-heavy WRF case such as WRF-Chem with many tracers transport this speed-up is significant.

LIST OF MODIFIED FILES:
dyn_em/module_advect_em.F

TESTS CONDUCTED:

Regression tests (WTF) PASS

Able to get bit-for-bit results for WTF tests between modified and original codes.

Top-hat advection tests from the kernel show the same results.