Modifications for reading ASCII observations with large numbers of levels (>1001) (#59)
TYPE: bug fix

KEYWORDS: WRFDA, ASCII observations, max_ob_levels, runtime error, write format

SOURCE: Internal

DESCRIPTION OF CHANGES:

A user was attempting to assimilate some very detailed soundings with more than 2000 vertical levels. This was hitting a hard-coded maximum "max_ob_levels=1001", which is intended to print a warning message and then ignore all levels past the first 1001. However, there was a runtime error when code block containing the warning message was reached, which presented in strange ways (gfortran claimed it was reaching the "End of file", ifort claimed "too many records in I/O statement"). I traced this down to the inclusion of slashes (/) in the fmt specification of this write statement. The inital fix was to remove the slashes, which allowed the code to run normally.

When testing this fix, I found that assimilation of these large observations still failed in debug mode. This is because, in da_scan_obs_ascii.inc, the ASCII observations are read and stored in the "platform" structure. Each level is stored in "platform%each(i)", where "i" is the level number. platform%each is defined as having "max_ob_levels" (1001) levels, which is why the limit check in da_read_obs_ascii.inc exists. However, to avoid writing past the end of the array in da_scan_obs_ascii.inc, we need to add this check here as well. We add an if check to each do loop over the number of observation levels in these two subroutines, and move the warning message to da_scan_obs_ascii since that is the first instance of this exception.

LIST OF MODIFIED FILES:
M var/da/da_obs_io/da_read_obs_ascii.inc
M var/da/da_obs_io/da_scan_obs_ascii.inc

TESTS CONDUCTED: Ran test with a sounding observation with a greater amount of levels than max_ob_levels. This case failed at runtime originally, and succeeded with the fix (for both gnu and intel). WRFDA Regression tests passed as expected. Added new large observation to existing WRFDA regression tests to check for similar problems in the future; these fail before the test and succeed afterwards, even in debug mode.