splitting the PD advection limiter loop for better performance for WENOPD(#61)
### Splitting the advection PD loop for wenopd advection

__TYPE:__ Improvement

__KEYWORDS:__ Advection, Positive Definite,WENO, Optimization

__SOURCE:__ Negin Sobhani (UIowa) | Davide Del Vento (NCAR)

__DESCRIPTION OF CHANGES:__
This change is similar to pull request #46 for PD advection limiter loop but in wenopd advection subroutine. We splitted the pd advection limiter loop to improve vectorization of this loop. 

This change showed a 112% speed-up (2.12X) for PD condition section of the code (the loop that has been split) and of ~6% speed-up for the advection kernel for intel compiler (-O3).  For GNU compiler (-Ofast), we have 100% speed-up on the PD limiter section of the code and total of ~3.2% speed-up for the advection kernel. For PGI compiler with -O3, we have only 35% speed-up for the PD limiter loop.

These results are less than pull request #46, since this positive definite condition is responsible for less computation in wenopd compared to PD.
Similar to #46, depending on how much advection you have in your WRF case, this speed-up might be significant.

__LIST OF MODIFIED FILES:__  
dyn_em/module_advect_em.F

__TESTS CONDUCTED:__   
1)  Regression tests (WTF): PASS
2) Able to get bit-for-bit results for WTF tests between modified and original codes.
3) Top-hat advection tests from the kernel show the same results.