Moist potential temperature lateral boundary fixes: use_theta_m=1 (#53)
Moist potential temperature lateral boundary fixes: use_theta_m=1

TYPE: bug fix

KEYWORDS: use_theta_m, moist, theta, BC, boundary

SOURCE: Bug hunt prompted by Bill Gustafson (PNNL)

DESCRIPTION OF CHANGES:

There are two modifications that are fixing bugs in the same subroutine:
theta_and_thetam_lbc_only. This routine computes the lateral boundary tendencies
for potential temperature for specified or for nested boundary conditions. When
dry potential temperature comes in, the moist theta BCs are computed. Similarly,
when moist theta is input, the output is the tendencies for the dry theta.

The horizontal indexes for the computation of the lateral boundary condition for
potential temperature do NOT need to be anything other than the ususal its->ite, and
jts->jte. The machinations with +4 and -4, depending on OpenMP or MPI was all wrong.
The potential temperature field is on mass points, the mu-coupling is on mass points,
there is no averaging, so no need for any stencil at all. This should be a computation
only on a single column, and that is what the indexing now does.

Two fields, t_bdy and t_bdy_tend, were originally INOUT variables. They are both used on the
right hand side of the equal sign, and they are used on the LHS also. Because of
this usage, care needed to be taken that an updated value was not subsequently (and
incorrectly) used on the RHS of the equations. Unfortunately, that care was not taken.
What is done now is to hold on to the original values (orig_t_bdy, orig_t_bdy_tend), and then ahead of each loop that uses the arrays, assign the copies t_bdy and t_by_tend to be used on the RHS of the statements. When computed, the values are stored in new_t_bdy and new_t_bdy_tend. At the end of the routine, the new_t_bdy and new_t_bdy_tend are put into the orig_t_bdy and orig_t_bdy_tend arrays.

LIST OF MODIFIED FILES:

M dyn_em/module_bc_em.F

TESTS CONDUCTED:

Old regression test passes with mods
New extended regression test that has use_theta_m=1 passes for serial, OpenMP, and MPI