Remove HWRF/NMM variable (avgPchg) from the history file. (#115)
TYPE: bugfix

KEYWORDS: NMM, HWRF 

SOURCE: internal                      

DESCRIPTION OF CHANGES: This variable causes some bit-for-bit errors
for a few of the NMM regression test cases.  It is not used for non-HWRF NMM.  
This is a work-around to avoid wrfout file diffs.

LIST OF MODIFIED FILES:

    M       Registry/Registry.NMM

TESTS CONDUCTED:

WTF completes as expected.