Add new file (p3_lookup_table_1.dat) to those removed with "./clean -a" (#127)
TYPE: bug fix

KEYWORDS: clean, p3_lookup_table_1.dat

SOURCE: internal

DESCRIPTION OF CHANGES:
In the "./clean" script is a list of files that are linked or copied from the WRFV3/run directory into the various WRFV3/test/* directories.  Those linked and copied data files are removed from the WRFV3/test/* directories with the issue of the "./clean" command.  A new file has been added to the WRFV3/run, so it needs to be added to the list of files to be removed.

LIST OF MODIFIED FILES:
M       clean

TESTS CONDUCTED:
1. After a build, and then running the "./clean a" script, a "git status" shows no left-over data files in the test/em_real directory.