Fix 3-domain restart by allowing climatological ozone interpolated to each nest domains in RRTMG (#246)
TYPE: bug fix

KEYWORDS: ozone input, RRTMG

SOURCE: internal

DESCRIPTION OF CHANGES:

A restart problem using 3 domains was traced to using o3input = 2 option. This option interpolates climatological ozone onto coarse grid, and uses special mechanism (p2c) to interpolate it to the nest. This appears to work ok with 2 domain cases in restart, but fails 3-domain restart. The fix is to do interpolation directly onto each nest grids.

In this PR "problem" is defined as not being bit-for-bit when comparing a restart to the original run.

LIST OF MODIFIED FILES:

modified: phys/module_radiation_driver.F

TESTS CONDUCTED: Regtested with WTF-3.07, all pass. Tested on a case with and without this change on restart.